Core - Rename CefSharp.Internals.CefPostDataWrapper to CefSharp.PostData
- Made public
- Added public constructor

Related Issue #1721
